Bump mix.lock, format change only (extra hash)
On mix 1.10.2 and hex 0.20.5, when I run `mix deps.get` the lock file changes format:
there is a new hash at the end of each line (with the exception of `prometheus_process_collector`
which I see has a nested definition.

There are no other content changes, no new versions or anything.

Per elixirforum there is no spec on this file; I'm proposing I just commit the new one, and we react
if there is git churn on the file.

Closes #53