Set general tracer span attributes in separate module.
First attempt at abstracting some of the verbosity of
Tracer.set_attributes to a module def, rather than repeating
similar code in each def we wish to observe.

Currently only setting tracer spans in 2 resolver defs,
and only for votes (1 of 2 defs) and create_vote (1 of 3 defs).
Not sure we need to also observe the related core api defs, as
the data returned is very similar, and the resolver defs 'wrap'
the related core api defs.