Improve description of absinthe result query in schema.ex
Now reads '...voting method (optional)...' to highlight that the
parameter is not required.
The query will return a (handled) errror in the case that no such
record exists, including for queries without a votingMethod param
when no 'default' result exists, and querying with a votingMethod
param when no such votingMethod exists.