Use voting method in absinthe Votes query.
proposalUrl and votingMethods fields are optional, but
providing only a voting method, without a proposal url, will
result in an error explaining that a proposal url is
required IF a voting method is specified.
If a proposal url is specified, with no voting method, then
votes for the proposal url and voting method with name
"default" will be returned.
Providing neither will return all votes for an organization.
An empty list is returned if no matching result(s) found.